Вынесение делегатов в отдельный файл.
