+++ .gitignore
@@ -1,0 +1,1 @@
+HM12.xcodeproj/project.xcworkspace/xcuserdata/aliaksandrpiatrovich.xcuserdatad/UserInterfaceState.xcuserstate
